Fine-tuning report, 16 buttons

+++ app/include/zmk/mouse.h
@@ -9,5 +9,5 @@
 #include <zephyr.h>
 #include <dt-bindings/zmk/mouse.h>
 
-typedef uint8_t zmk_mouse_button_flags_t;
-typedef uint8_t zmk_mouse_button_t;
+typedef uint16_t zmk_mouse_button_flags_t;
+typedef uint16_t zmk_mouse_button_t;
